gdprApplies, refactoring, fix tests
