+++ libraries/intentIqUtils/storageUtils.js
@@ -1,7 +1,8 @@
-import {logError} from '../../src/utils.js';
-import {SUPPORTED_TYPES} from '../../libraries/intentIqConstants/intentIqConstants.js';
+import {logError, logInfo} from '../../src/utils.js';
+import {SUPPORTED_TYPES, FIRST_PARTY_KEY} from '../../libraries/intentIqConstants/intentIqConstants.js';
 
 const MODULE_NAME = 'intentIqId';
+const PCID_EXPIRY = 365;
 
 /**
  * Read data from local storage or cookie based on allowed storage types.
@@ -21,6 +22,35 @@ export function readData(key, allowedStorage, storage) {
     logError(`${MODULE_NAME}: Error reading data`, error);
   }
   return null;
+}
+
+/**
+ * Store Intent IQ data in cookie, local storage or both of them
+ * expiration date: 365 days
+ * @param {string} key - The key under which the data will be stored.
+ * @param {string} value - The value to be stored (e.g., IntentIQ ID).
+ * @param {Array} allowedStorage - An array of allowed storage types: 'html5' for Local Storage and/or 'cookie' for Cookies.
+ * @param {Object} storage - The storage handler object that manages Local Storage and Cookies.
+ * @param {Object} firstPartyData - Contains user consent data; if isOptedOut is true, data will not be stored (except for FIRST_PARTY_KEY).
+ */
+export function storeData(key, value, allowedStorage, storage, firstPartyData) {
+  try {
+    if (firstPartyData?.isOptedOut && key !== FIRST_PARTY_KEY) {
+      return;
+    }
+    logInfo(MODULE_NAME + ': storing data: key=' + key + ' value=' + value);
+    if (value) {
+      if (storage.hasLocalStorage() && allowedStorage.includes('html5')) {
+        storage.setDataInLocalStorage(key, value);
+      }
+      if (storage.cookiesAreEnabled() && allowedStorage.includes('cookie')) {
+        const expiresStr = (new Date(Date.now() + (PCID_EXPIRY * (60 * 60 * 24 * 1000)))).toUTCString();
+        storage.setCookie(key, value, expiresStr, 'LAX');
+      }
+    }
+  } catch (error) {
+    logError(error);
+  }
 }
 
 /**
